use React Provider for convoListItem
this is to avoid passing down the prop to all the components
