+++ ts/hooks/useParamSelector.ts
@@ -3,13 +3,9 @@ import { PubKey } from '../session/types';
 import { UserUtils } from '../session/utils';
 import { StateType } from '../state/reducer';
 
-export function useAvatarPath(pubkey: string | undefined) {
-  return useSelector((state: StateType) => {
-    if (!pubkey) {
-      return null;
-    }
-    return state.conversations.conversationLookup[pubkey]?.avatarPath || null;
-  });
+export function useAvatarPath(convoId: string | undefined) {
+  const convoProps = useConversationPropsById(convoId);
+  return convoProps?.avatarPath || null;
 }
 
 export function useOurAvatarPath() {
@@ -20,39 +16,25 @@ export function useOurAvatarPath() {
  *
  * @returns convo.profileName || convo.name || convo.id or undefined if the convo is not found
  */
-export function useConversationUsername(pubkey?: string) {
-  return useSelector((state: StateType) => {
-    if (!pubkey) {
-      return undefined;
-    }
-    const convo = state.conversations.conversationLookup[pubkey];
-    if (!convo) {
-      return pubkey;
-    }
-    return convo?.profileName || convo?.name || convo.id;
-  });
+export function useConversationUsername(convoId?: string) {
+  const convoProps = useConversationPropsById(convoId);
+
+  return convoProps?.profileName || convoProps?.name || convoId;
 }
 
 /**
  * Returns either the nickname, profileName, or the shorten pubkey
  */
-export function useConversationUsernameOrShorten(pubkey?: string) {
-  return useSelector((state: StateType) => {
-    if (!pubkey) {
-      return undefined;
-    }
-    const convo = state.conversations.conversationLookup[pubkey];
-    if (!convo) {
-      return PubKey.shorten(pubkey);
-    }
-    return convo?.profileName || convo?.name || PubKey.shorten(convo.id);
-  });
+export function useConversationUsernameOrShorten(convoId?: string) {
+  const convoProps = useConversationPropsById(convoId);
+
+  return convoProps?.profileName || convoProps?.name || (convoId && PubKey.shorten(convoId));
 }
 
 /**
- * Returns either the nickname, profileName, or the shorten of the pubkeys given
+ * Returns either the nickname, the profileName, in '"' or the full pubkeys given
  */
-export function useConversationsUsernameOrFull(pubkeys: Array<string>) {
+export function useConversationsUsernameWithQuoteOrFullPubkey(pubkeys: Array<string>) {
   return useSelector((state: StateType) => {
     return pubkeys.map(pubkey => {
       if (pubkey === UserUtils.getOurPubKeyStrFromCache() || pubkey.toLowerCase() === 'you') {
@@ -73,16 +55,18 @@ export function useIsMe(pubkey?: string) {
 }
 
 export function useIsClosedGroup(convoId?: string) {
-  return useSelector((state: StateType) => {
-    if (!convoId) {
-      return false;
-    }
-    const convo = state.conversations.conversationLookup[convoId];
-    if (!convo) {
-      return false;
-    }
-    return (convo.isGroup && !convo.isPublic) || false;
-  });
+  const convoProps = useConversationPropsById(convoId);
+  return (convoProps && convoProps.isGroup && !convoProps.isPublic) || false;
+}
+
+export function useIsPrivate(convoId?: string) {
+  const convoProps = useConversationPropsById(convoId);
+  return Boolean(convoProps && convoProps.isPrivate);
+}
+
+export function useIsPinned(convoId?: string) {
+  const convoProps = useConversationPropsById(convoId);
+  return Boolean(convoProps && convoProps.isPinned);
 }
 
 export function useConversationPropsById(convoId?: string) {
